delete.rb: Add command to delete jira tickets
Users should be able to delete jira tickets as well as
create them. With "delete" command, users can delete
a ticket as well as tickets associated with sub-tasks.

Part of the command maintains your git branch by deleting
the ticket from it. As a result, users are prompted to
create then select a new branch to switch.