Automatically renew cookie when connect fails.
If jira-cli fails to connect to the configured Jira instance
specified in .jira-cli, then jira-cli will provide a user
the option to renew the cookie session. It does so by asking
for a password and then posting for the cookie session.

This option only occurs if .jira-cli was configured for a
previous cookie session.